Break out utilities methods of MapLayout into a Mixin class.
